requirements.txt
asgiref==3.5.2
Django==4.0.6
sqlparse==0.4.2

djangorestframework~=3.14.0